+++ components/float-button/FloatButton.tsx
@@ -11,7 +11,7 @@ import Badge from '../badge';
 import type { BadgeProps } from '../badge';
 import Button from '../button';
 import type { ButtonHTMLType } from '../button';
-import type { ButtonSemanticName } from '../button/Button-';
+import type { ButtonSemanticName } from '../button/Button';
 import { ConfigContext } from '../config-provider';
 import useCSSVarCls from '../config-provider/hooks/useCSSVarCls';
 import Tooltip from '../tooltip';
